Add raw on-disk video cache.
This is useful for faster testing iterations.
It avoids waiting for the footage to decode on startup.
Using file memory mapping, we let the OS handle caching.
The OS disk cache also has the advantage of being kept in memory between runs.
A cache is only created for shot < 1GiB (e.g 17s of 1080@30).