Add a new hybrid region tracker for motion tracking to libmv, and add it as an option (under "Hybrid") in the tracking settings. The region tracker is a combination of brute force tracking for coarse alignment, then refinement with the ESM/KLT algorithm already in libmv that gives excellent subpixel precision (typically 1/50'th of a pixel)
This also adds a new "brute force" region tracker which does a
brute force search through every pixel position in the destination
for the pattern in the first frame. It leverages SSE if available,
similar to the SAD tracker, to do this quickly. Currently it does
some unnecessary conversions to/from floating point that will get
fixed later.

The hybrid tracker glues the two trackers (brute & ESM) together
to get an overall better tracker. The algorithm is simple:

1. Track from frame 1 to frame 2 with the brute force tracker.
    This tries every possible pixel position for the pattern from
    frame 1 in frame 2. The position with the smallest
    sum-of-absolute-differences is chosen. By definition, this
    position is only accurate up to 1 pixel or so.
2. Using the result from 1, initialize a track with ESM. This does
    a least-squares fit with subpixel precision.
3. If the ESM shift was more than 2 pixels, report failure.
4. If the ESM track shifted less than 2 pixels, then the track is
    good and we're done. The rationale here is that if the
    refinement stage shifts more than 1 pixel, then the brute force
    result likely found some random position that's not a good fit.