Added solver for modal camera motion, currently supports only tripod solving
This solver is intended to deal with such camera motions as tripod and panning,
where it's impossible to reconstruct exact position of markers in 3d view.

It projects markers onto sphere and uses rigid registration of rotation to
find rotation angles which makes bundles from previous and current frame be
as closest as it's possible.