Camera Tracking: allow fallback to reprojection resection by user demand
This fixes some "regressions" introduced in previous commit
which lead to much worse solution in some cases. Now it's
possible to bring old behavior back.

Perhaps it's more like temporal solution for time being smarter
solution is found. But finding such a solution isn't so fast,
so let's bring manual control over reprojection usage.

But anyway, imo it's now nice to have a structure which could
be used to pass different settings to the solver.