Automatic keyframe selection based on Pollefeys's criteria
This commit implements automatic keyframe selection algorithm
based on Pollefeys's criteria (F-GRIC is smaller than H-GRIC
and correspondence ratio is more then 90%).

It is implemented as a part of simple pipeline and returns
vector of keyframe images for a given Tracks structure.

For simple pipeline reconstruction two best keyframes are
expected to be selected from all detected candidates.
Criteria for this selection could be reprojection error of
solution from two candidate keyfames.

Unfortunately, it's not fully workable yet, hopefully would
be fixed soon.